# eager load analysis results for quicker data load: