--- a/.gitignore
+++ b/.gitignore
@@ -18,3 +18,4 @@ docs/.*
 _ide_helper.php
 *.DS_Store
 /database/views/samples_merged*
+.idea/*