After pushing the optimisation to move to ajax loading for samples, I realised it was still taking a long time to load. Realised we had an "n+1" issue for database queries (each sample was getting multiple db queries to load the individual analysis results, meaning we had (sample_count * analysis_tables) queries happening for each page load). 

This is now updated:
- eager loading analysis results in the getJson() method to avoid n+1 calls to database;
- in the Model, the attribute getters now use pre-loaded collections instead of relationship methods (to avoid redoing database queries):

e.g. 
- `$this->analysis_poxc()->first()` uses the relationship method, which means a new database call; 
- `$this->analysis_poxc->first()` will check if the analysis_poxc collection already exists, and only go to the database if it doesn't.