@dan-tang-ssd 2022-09-28
Thank you Dave for the performance finetuning patch.
As there are not too many records in my local env, I believe this patch can greatly improve performance with large sample data.
Below please find my testing result FYI.

---

Soils Platform – PR 96 Testing

Perform testing in local env BEFORE and AFTER getting latest code

BEFORE getting latest code

First project with 3 attempts: 717ms, 605ms, 597ms
First project average time for JSON response: (717 + 605 + 597) / 3 = 639.67ms

Second project with 3 attempts: 1.93s, 1.87s, 1.97s
Second project average time for JSON response: (1.93 + 1.87 + 1.97) / 3 = 1.92s

AFTER getting latest code

First project with 3 attempts: 571ms, 557ms, 544ms
First project average time for JSON response: (571 + 557 + 544) / 3 = 557.33ms

Second project with 3 attempts: 709ms, 710ms, 714ms
Second project average time for JSON response: (709 + 710 + 714) / 3 = 711ms


Testing result with screen shots can be found in below document:
[Soils_PR_96_Testing.docx](https://github.com/stats4sd/ccrp-soils/files/9665133/Soils_PR_96_Testing.docx)