@dave-mills 2022-09-28
Nice - thanks! Good to know it's a decent performance boost even with small data. It's definitely a speed boost for larger datasets (I seeded my local with ~600 samples).